resource files/docker/docker-compose-flask-nginx-postgreSql/app/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.1